Move pod fast delete test cases to the common suite
